+++ ftmgram/__init__.py
@@ -16,7 +16,7 @@
 #  You should have received a copy of the GNU Lesser General Public License
 #  along with Pyrogram.  If not, see <http://www.gnu.org/licenses/>.
 
-__version__ = "3.3.0"
+__version__ = "3.5.0"
 __license__ = "GNU Lesser General Public License v3.0 (LGPL-3.0)"
 __copyright__ = "Copyright (C) 2024-present FTM DEVELOPERZ <https://github.com/ftmdevz> <ftmdeveloperz@gmail.com>"
 
@@ -33,6 +33,7 @@ class ContinuePropagation(StopAsyncIteration):
     pass
 
 
-from . import raw, types, filters, handlers, enums
+from . import raw, types, filters, handlers, enums, helpers
 from .client import Client
 from .sync import idle, compose
+