+++ ftmgram/methods/chats/__init__.py
@@ -40,7 +40,9 @@ from .edit_forum_topic import EditForumTopic
 from .delete_message_reaction import DeleteMessageReaction
 from .create_folder_invite_link import CreateFolderInviteLink
 from .get_chat import GetChat
+from .get_bot_chat_history import GetBotChatHistory
 from .get_chats_for_folder_invite_link import GetChatsForFolderInviteLink
+
 from .get_chat_event_log import GetChatEventLog
 from .get_folder_invite_links import GetFolderInviteLinks
 from .get_chat_member import GetChatMember
@@ -212,6 +214,8 @@ class Chats(
     SetChatProtectedContent,
     UnpinAllForumTopicMessages,
     PurgeMessages,
+    GetBotChatHistory,
 ):
     pass
+